MaskGIT/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aiofiles=22.1.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
aiosqlite=0.19.0=pypi_0
albumentations=1.3.0=pypi_0
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.5.7=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=py311h5eee18b_3
chardet=4.0.0=py311h06a4308_1003
charset-normalizer=3.1.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.3=pypi_0
contourpy=1.0.5=py311hdb19cb5_0
cryptography=39.0.1=py311h9ce1e76_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
einops=0.6.1=pypi_0
executing=1.2.0=pypi_0
expat=2.2.10=h9c3ff4c_0
fastjsonschema=2.16.3=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py311h06a4308_0
flatbuffers=23.5.26=pypi_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.25.0=pyhd3eb1b0_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.5.0=pyh1a96a4e_0
gast=0.4.0=pypi_0
gdown=4.7.1=pypi_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
google-auth=2.19.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.54.2=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.8.0=pypi_0
icu=58.2=hf484d3e_1000
idna=3.4=py311h06a4308_0
imageio=2.28.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.23.0=pypi_0
ipython=8.13.2=pypi_0
ipython-genutils=0.2.0=pypi_0
isoduration=20.11.0=pypi_0
jax=0.4.11=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=py311h06a4308_0
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
json5=0.9.11=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter-client=8.2.0=pypi_0
jupyter-core=5.3.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.5.0=pypi_0
jupyter-server-fileid=0.9.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyter-server-ydoc=0.8.0=pypi_0
jupyter-ydoc=0.2.4=pypi_0
jupyterlab=3.6.3=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.22.1=pypi_0
keras=2.12.0=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py311h6a678d5_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lazy-loader=0.2=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libclang=16.0.0=pypi_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=he513fc3_3
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lightning-utilities=0.8.0=pyhd8ed1ab_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=py311h38be061_0
matplotlib-base=3.7.1=py311ha02d727_1
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.6=py311ha02d727_1
mkl_random=1.2.2=py311ha02d727_1
ml-dtypes=0.2.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py311h06a4308_0
multidict=6.0.4=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.4.0=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py311h06a4308_1
notebook=6.5.4=pypi_0
notebook-shim=0.2.3=pypi_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.23.5=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pyhd8ed1ab_0
opencv-python-headless=4.7.0.72=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=23.1=pyhd8ed1ab_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py311h6a678d5_0
pip=23.1.2=pypi_0
platformdirs=3.5.0=pypi_0
ply=3.11=py_1
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.23.2=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pypi_0
pynvml=11.5.0=pypi_0
pyopenssl=23.0.0=py311h06a4308_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py311h6a678d5_0
pyqt5-sip=12.11.0=py311h6a678d5_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.3=h7a1cb2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-json-logger=2.0.7=pypi_0
python_abi=3.11=2_cp311
pytorch=2.0.1=py3.11_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-fid=0.3.0=pypi_0
pytorch-lightning=2.0.2=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
pyyaml=6.0=py311h5eee18b_1
pyzmq=25.0.2=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
requests=2.30.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
send2trash=1.8.2=pypi_0
setuptools=66.0.0=py311h06a4308_0
sip=6.6.2=py311h6a678d5_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.2=pypi_0
sympy=1.12=pyh04b8f61_3
tbb=2021.8.0=hdb19cb5_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=2.0.2=py311_cu117
torchmetrics=0.11.4=pyhd8ed1ab_0
torchsummary=1.5.1=pypi_0
torchtriton=2.0.0=py311
torchvision=0.15.2=py311_cu117
tornado=6.3.1=pypi_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.9.0=pypi_0
typing_extensions=4.5.0=py311h06a4308_0
tzdata=2023c=h04d1e81_0
uri-template=1.2.0=pypi_0
urllib3=1.26.15=py311h06a4308_0
utils=1.0.1=pypi_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.3.5=pypi_0
wheel=0.38.4=py311h06a4308_0
wrapt=1.14.1=pypi_0
xz=5.4.2=h5eee18b_0
y-py=0.5.9=pypi_0
yaml=0.2.5=h7f98852_2
yarl=1.9.2=pypi_0
ypy-websocket=0.8.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
